Refactor core components’s folder structure
- Move App, Root, and Routes components to src/components/core
- Update paths to import statements
